+++ CHANGELOG.md
@@ -80,6 +80,7 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
   - This will allow image scans to reflect the latest features (for example, we support a new language, we will re-index an image to see if artifacts of the new language exist)
   - This will also clean up manifests/Index Reports from Scanner V4 DB which are no longer relevant in the environment
   - This will mean Scanner V4 Indexer will now pull images from the registry more than just once, putting some more load on registries.
+  - Scanner V4 Indexer periodically checks for expired manifests at the interval specified by `ROX_SCANNER_V4_MANIFEST_GC_INTERVAL` (current default: 6 hours).
 
 ## [4.5.0]